Rename Float->Decimal and Bool->Binary across all user-facing UI text
Labels only: dropdown option text, list badges (now via shared
typeDisplayName helper), node descriptions, validation hints, Help tab,
README, NODES_REFERENCE. Internal type ids ('bool'/'float' in schema,
configs, compilers, .gcaproj) are untouched - zero behavior change.

+++ src/modeler/panels/PropertiesPanelContent.tsx
@@ -645,7 +645,7 @@ export function PropertiesPanelContent() {
                         </div>
                         {floatFreqDisabled && (
                           <span style={{ color: '#e0a050', fontSize: '0.62rem', fontStyle: 'italic', paddingLeft: 4 }}>
-                            Float-binned frequency categories depend on runtime range. Change this indicator&apos;s aggregation to Total, or pick a different indicator.
+                            Decimal-binned frequency categories depend on runtime range. Change this indicator&apos;s aggregation to Total, or pick a different indicator.
                           </span>
                         )}
                       </div>
